feat: Total directory structure overhaul, more...
Removed K8s. Something about the way networking works is currently above my skill works, and I do not think that I have the base knowledge of K8s to deploy it in such a declarative way - yet. For now, I will instead use `oci-containers`, as they are unbelievably easy to deploy!

I also redid the structure of the various systems, as I was starting to lose track of changes made on one system and how they affected another.